test/allocator: move bluestore allocator's benchmarks to a standalone UT
(cherry picked from commit ac2ec80cddc51f1642849c7bf226e9bed116c0d6)
